backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
curlify==2.2.1
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
facebook-business==12.0.1
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pycountry==22.1.10
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
yarl==1.7.2
